refactor(NewArticle): colocate calculateTotalArticles function in the useGetArticles hook
