+++ src/hooks/useGetArticles.ts
@@ -1,16 +1,17 @@
-import { ArticleCodec } from '@/types'
-import { fetcher, type ValidationError, type UnknownError } from '@/libs'
 import {
   useInfiniteQuery,
   type QueryFunctionContext,
   type QueryKey,
   type UseInfiniteQueryOptions,
 } from '@tanstack/react-query'
+import { ArticleCodec } from '@/types'
+import { fetcher, type ValidationError, type UnknownError } from '@/libs'
 import { type Either } from 'fp-ts/Either'
 import { withMessage } from 'io-ts-types'
 import { isSome, fromNullable } from 'fp-ts/Option'
-import { calculateTotalArticles } from '@/libs/calculateTotalArticles'
+import { match } from 'fp-ts/Either'
 import { type, array, number, type TypeOf, type OutputOf } from 'io-ts'
+import { pipe } from 'fp-ts/function'
 
 export type QueryFiltersProps = {
   limit?: string
@@ -37,6 +38,30 @@ export const defaultArticlesLimit = 6
 
 export const defaultFilters = {
   limit: defaultArticlesLimit.toString(),
+}
+
+export const calculateTotalArticles = (
+  lastPage: Either<ValidationError, GetArticlesOutput>
+) => {
+  const getLimit = (totalLastFetch: number) =>
+    totalLastFetch + defaultArticlesLimit
+
+  const getFilters = (limit: number): QueryFiltersProps => ({
+    limit: limit.toString(),
+  })
+
+  const getTotalArticles = (articles: GetArticlesOutput) => {
+    const totalLastFetch = articles.articles.length
+    const total = articles.articlesCount
+    const isEqual = totalLastFetch === total
+    return isEqual ? null : pipe(totalLastFetch, getLimit, getFilters)
+  }
+
+  const onRight = (articles: GetArticlesOutput) => getTotalArticles(articles)
+
+  const onLeft = () => null
+
+  return pipe(lastPage, match(onLeft, onRight))
 }
 
 const oneMinute = 60 * 1000
