+++ src/libs/index.ts
@@ -1,4 +1,3 @@
-export * from '@/libs/calculateTotalArticles'
 export * from '@/libs/errors'
 export * from '@/libs/expression'
 export * from '@/libs/fetcher'
